devopsloft/spinner/requirements.txt
cffi==1.14.1
Click==7.1.2
python-dotenv==0.14.0
j2cli==0.3.10
docker-compose==1.26.2
boto3==1.14.42